requirements.txt
hermes_python>=0.1
requests
paho-mqtt
